Translate recorded objects into the McMini model
This commit begins the process of transfering the multithreaded
primitives (e.g. mutexes, threads, etc.)  recorded prior to restart by
`libmcmini.so`. The process essentially entails inspecting the fields of
the structs used during recording and creating the appropriate object.
For transitions, we simply inspect the shared memory and call the
appropriate callback function just as if we were running McMini in classic
model checking mode. The result is that we can think of `libmcmini.so`
handing McMini objects as McMini "observing" those objects for the first
time, just as would happen at runtime in classic model checking mode.