Record runnable objects befor pending operations
Here we address a more subtle bug translating the recording of
`libmcmini.so` into the McMini model.  Some transitions, such as
`pthread_join()`, need the state of runnable model objects to be described
adequately. However, we were only recording the state of non-runnable
objects (e.g. mutexes) prior to determining the "next steps" for each
thread.

This commit resolves this issue by also recording the state of these
runnable objects prior to the second phase.