Fix (again) improper FIFO use
This commit resolves a bug using FIFOs. The trouble was that the template
thread in the restarted process was using

```
open("/tmp/mcmini", O_CREAT | O_WRONLY)
```

instead of

```
mkfifo("/tmp/mcmini", S_IRUSR | S_IWUSR)
open("/tmp/mcmini", O_WRONLY)
```

This led to a race between the McMini process and the template thread. If
McMini called `mkfifo` before the `open` call in the template thread,
the code would succeed.  If the template thread called `open()` before
the McMini process, and if a mkfifo didn't already exist, the new file
was just a regular file with erroneous permission bits.

The solution is simply to call `mkfifo()` in both processes.  Only one
process will succeed, but that's ok. `mkfifo` returns with EEXIST in the
process which fails. Since the FIFO now exists and is correctly opened
on each side, the transfer succeeds.