Use `dmtcp_env.txt` at runtime to influence env vars
McMini now passes the `--modify-env` flag to `dmtcp_launch` at
record-time. The toruble is that `dmtcp_restart` by default will
(reasonably) restore the state of the environment as it looked prior to
checkpoint. For programs which wish to execute in a new environment,
however, the `modify_env` DMTCP plugin must be loaded prior to
checkpoint. This is done using the aforementioned flag.

A new `dmtcp_target` subclassing `target` was also added. The difference
here is that `dmtcp_target` will at runtime open the `dmtcp_env.txt`
file in the directory of the checkpoint file. When new environment
variables are set at runtime by McMini, the file is truncated and
refreshed. If the environment variables are constant, however, the file
will not repeatedly be opened/truncated and rewritten.

This seems to have a noticeable effect on the runtime performance of
McMini which should be investigated in the future. The multithreaded
fork implementation should address this however.