Support SIGSEGV inside template and branch processes
This commit introduces support for assertion failures and segmentation
faults inside the branch process or inside a template process during
deep debugging. The challenge in supporting detecting these errors is
that errors in branch processes are seen inside the template process and
are not reported directly to the `mcmini` process because the `mcmini`
process is a _grandparent_ of the branch process. We attempted to
circumvent this using `prctl(PR_SETSUBREAPER, 1)` which ensures that all
descendants of the given process become the de facto parent when the
parent of any descendant dies. However, since the template process, the
parent of the branch process, does not die, it's the template process
which receives a SIGCHLD when the branch process exits (unexpectedly or
expectedly). The template therefore must then notify the `mcmini`
process that something went wrong. We do this by using the `kill(2)`
system call inside the SIGCHLD handler of the template process.
Fortunately, the `kill(2)` system call is async-signal-safe, so it's
safe to use in signal handlers