Add additional sem_wait/sem_post rendez-vous
A `sem_post()` was getting mixed up between two multithreaded_fork()
calls (one immediately following the other). Essentially, a child thread
in the template process appeared to not have exited its
SIG_MULTITHREADED_FORK handler before the second multithreaded_fork()
call came in. The child thread would have been calling `sem_wait()`.
We suspect the bug occurs because we receive a signal inside the signal
handler while calling `sem_wait()`, which is known not to be
async-signal-safe. A workaround is to use the self-pipe trick instead of
sem_wait. For now, we add an additional sem_post/sem_wait such that a
signal sent via tgkill during the next multithreaded_fork call can only
occur when a child thread is calling `sem_post()`, which is
async-signal-safe.