Add `fast_multithreaded_fork()` impl
The existing `multithreaded_fork()` implementation does not assume a
particular location in which it will be called: userspace threads are
assumed to be possibly executing in different locations in the code.
However, in deep debugging, userspace threads will always be executing
in a known location: `thread_handle_after_dmtcp_restart()`. Using this,
we can amortize the cost of getcontext() and writing the pthread
descriptors into a global storage at the "getting into a stable state
phase". The idea is that the context is all that really matters to
successfully execute multithreaded_fork().

With this knowledge in mind, the `fast_multithreaded_fork()` is
introduced which simply reads from global memory the appropriate
contexts inside the child process. Otherwise, it is conceptually very
close to fork().