Record `pthread_t` mapping for pthread_join() for more stability
Raw `pthread_t` structs were passed in shared memory between the McMini
process and the branch processes after multithreaded_fork(). However,
this is suspect because although the `pthread_t` struct is patched
between `multithreaded_fork()` calls, it may not be exactly "equal"
copy-wise. Instead, we now keep track of a light-weight mapping of
pthread_t to runner IDs. The higher-level goal is to be able to track
objects persistently across traces despite the fact that these objects
will appear in (hundreds) of traces.