Fix deadlock logic issue
Detecting deadlocks is a key feature of McMini. An edge case in deadlock
detection occurs when every thread in the program has exited. In such
cases, the program trivially can't perform any new actions, so we don't
consider it to be in a deadlock. However, the logic for detecting this
was incorrect prior to this commit since we claimed that a deadlock was
impossible so long as a _single_ thread had exited.